TCVP-3144 change statute service to use v2 lookup

+++ src/backend/TrafficCourts/TrafficCourts.OrdsDataService/Justin/IStatuteRepository.cs
@@ -3,6 +3,6 @@ namespace TrafficCourts.OrdsDataService.Justin;
 
 public interface IStatuteRepository
 {
-    Task<Statute?> GetAsync(decimal stat_id, CancellationToken cancellationToken);
+    Task<Statute?> GetAsync(int stat_id, CancellationToken cancellationToken);
     Task<List<Statute>> GetListAsync(CancellationToken cancellationToken);
 }
